+++ .gitignore
@@ -4,3 +4,4 @@
 /local-testing-convos/
 /local-testing-configs/
 /state
+__pyacache__/
