+++ config.json
@@ -1,6 +1,6 @@
 {
   "command_channels": [
-    1204624478496227391
+    1192895746672951340
   ],
   "defaults": {
     "engine": "gpt-4-turbo-preview",
