+++ myconfig.json
@@ -1,0 +1,28 @@
+{
+    "command_channels": [
+      1204624478496227391
+    ],
+    "defaults": {
+      "engine": "gpt-4-turbo-preview",
+      "timeout": 600
+    },
+    "channels": [
+      {
+        "name": "duck-pond",
+        "prompt_file": "prompts/standard-rubber-duck.txt"
+      },
+      {
+        "name": "other-pond",
+        "prompt_file": "prompts/standard-rubber-duck.txt",
+        "engine": "gpt-3.5-turbo-16k"
+      },
+      {
+        "name": "gpt4",
+        "engine": "gpt-4-turbo-preview"
+      },
+      {
+        "name": "gpt-3.5-turbo",
+        "engine": "gpt-3.5-turbo"
+      }
+    ]
+  }
